requirements.txt
PyGithub
python-dotenv
pandas
numpy
scikit-learn
xgboost
tqdm
joblib
sentence-transformers
matplotlib
